+++ Examples/Io/Csv/CMakeLists.txt
@@ -14,7 +14,7 @@ add_library(
   src/CsvTrackingGeometryWriter.cpp
   src/CsvTrackParameterReader.cpp
   src/CsvTrackParameterWriter.cpp
-  src/CsvMultiTrajectoryWriter.cpp 
+  src/CsvTrackWriter.cpp 
   src/CsvProtoTrackWriter.cpp
   src/CsvSpacePointWriter.cpp
   src/CsvBFieldWriter.cpp)
